+++ Editor/Editors/WorldEditor/WorldEditorView.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using Editor.Components;
+using Editor.Content;
 using Editor.GameCode;
 using System.Windows;
 using System.Windows.Controls;